get temp

+++ src/main/java/com/example/vibon/Model/Temperatura.java
@@ -14,10 +14,10 @@ public class Temperatura {
     @Temporal(TemporalType.TIMESTAMP)
     private Date dataCriacao;
 
-    private float temp;
+    private double temp;
 
 
-    public Temperatura(float temp){
+    public Temperatura(double temp){
         setCreationDateTime();
         this.temp = temp;
     }
